breaking: convert @cypress/grep to TypeScript and change what is exported
BREAKING CHANGE: @cypress/grep now exports a register function that can be imported from the users spec file. Additionally, the plugin can be imported via @cypress/grep/plugin.
