+++ .circleci/src/workflows/@workflows.yml
@@ -2368,9 +2368,34 @@ jobs:
     resource_class: small
     steps:
       - restore_cached_workspace
+      - run: yarn workspace @cypress/grep cypress:run:tags:before
+      - run: yarn workspace @cypress/grep cypress:run:tags:describe
+      - run: yarn workspace @cypress/grep cypress:run:tags:config
+      - run: yarn workspace @cypress/grep cypress:run:burn
+      - run: yarn workspace @cypress/grep cypress:run:tags:inherit
+      - run: yarn workspace @cypress/grep cypress:run:multiple-registrations
+      - run: yarn workspace @cypress/grep cypress:run:tags:nested-describe
+      - run: yarn workspace @cypress/grep cypress:run:skip
+      - run: yarn workspace @cypress/grep cypress:run:this
+      - run: yarn workspace @cypress/grep cypress:run:tags
+      - run: yarn workspace @cypress/grep cypress:run:each
+      - run: yarn workspace @cypress/grep cypress:run:grep
+      - run: yarn workspace @cypress/grep cypress:run:tags:or
+      - run: yarn workspace @cypress/grep cypress:run:tags:and
+      - run: yarn workspace @cypress/grep cypress:run:tags:and:not
+      - run: yarn workspace @cypress/grep cypress:run:or
+      - run: yarn workspace @cypress/grep cypress:run:and
+      - run: yarn workspace @cypress/grep cypress:run:and:not
+      - run: yarn workspace @cypress/grep cypress:run:filter:specs
+      - run: yarn workspace @cypress/grep cypress:run:tags:filter:specs
+      - run: yarn workspace @cypress/grep cypress:run:omit:specs
+      - run: yarn workspace @cypress/grep cypress:run:tags:omit:specs
+      - run: yarn workspace @cypress/grep cypress:run:untagged
+      - run: yarn workspace @cypress/grep cypress:run:tag:inverted
+      - run: yarn workspace @cypress/grep cypress:run:inverted
       - run:
-          name: Run tests
-          command: yarn workspace @cypress/grep cy:run
+          name: Run unit tests
+          command: yarn workspace @cypress/grep test
       - sanitize-verify-and-store-mocha-results:
           testResultsPath: npm/grep/test_results
