+++ npm/grep/cypress/e2e/burn.cy.ts
@@ -1,5 +1,3 @@
-/// <reference types="cypress" />
-
 // if we specify just the burn parameter
 // then this test will be repeated N times
 describe('burning a test N times', () => {
