+++ npm/grep/cypress/e2e/describe-tags.cy.ts
@@ -1,7 +1,3 @@
-/// <reference path="../../src/index.d.ts" />
-
-// @ts-check
-
 describe('block with no tags', () => {
   it('inside describe 1', () => {})
