+++ npm/grep/cypress/e2e/each.cy.ts
@@ -1,5 +1,3 @@
-/// <reference types="cypress" />
-
 // https://github.com/bahmutov/cypress-each
 import 'cypress-each'
