+++ npm/grep/cypress/e2e/inherits-tag.cy.ts
@@ -1,5 +1,3 @@
-/// <reference types="cypress" />
-
 describe('Screen A', { tags: ['@sanity', '@screen-a'] }, () => {
   it('loads', { tags: ['@screen-b'] }, () => {
     // do something that eventually sends the page to screen b.
