+++ npm/grep/cypress/e2e/multiple-registrations.cy.ts
@@ -1,10 +1,10 @@
-/// <reference types="cypress" />
-
 // register the plugin multiple times
 // to simulate including from support and spec files
 // https://github.com/cypress-io/cypress-grep/issues/59
-require('../../src/support')()
-require('../../src/support')()
-require('../../src/support')()
+const { register } = require('../../src/register')
+
+register()
+register()
+register()
 
 it('hello world', () => {})
