+++ npm/grep/cypress/e2e/nested-describe.cy.ts
@@ -1,6 +1,3 @@
-/// <reference path="../../src/index.d.ts" />
-
-// @ts-check
 describe('grand', () => {
   context('outer', { tags: '@smoke' }, () => {
     describe('inner', () => {
