+++ npm/grep/cypress/e2e/skip.cy.ts
@@ -1,4 +1,3 @@
-/// <reference types="cypress" />
 describe('tests that use .skip', () => {
   // use a template literal
   it(`works`, () => {})
