+++ npm/grep/cypress/e2e/tags/test1.cy.ts
@@ -1,5 +1,3 @@
-/// <reference types="cypress" />
-
 it('Test 1', { tags: ['smoke', 'regression'] }, () => {
   expect(true).to.be.true
 })
