+++ npm/grep/cypress/e2e/tags/test2.cy.ts
@@ -1,5 +1,3 @@
-/// <reference types="cypress" />
-
 it('Test 2', { tags: ['high', 'smoke'] }, () => {
   expect(true).to.be.true
 })
