+++ npm/grep/cypress/e2e/tags/test3.cy.ts
@@ -1,5 +1,3 @@
-/// <reference types="cypress" />
-
 it('Test 3', { tags: ['smoke'] }, () => {
   expect(true).to.be.true
 })
