+++ npm/grep/package.json
@@ -2,10 +2,36 @@
   "name": "@cypress/grep",
   "version": "0.0.0-development",
   "description": "Filter tests using substring",
-  "main": "src/support.js",
   "scripts": {
-    "cy:open": "node ../../scripts/cypress.js open --e2e -b electron --config specPattern='**/unit.js'",
-    "cy:run": "node ../../scripts/cypress.js run --config specPattern='**/unit.js'",
+    "build": "tsc",
+    "cypress:open": "node ../../scripts/cypress.js open",
+    "cypress:run:tags:before": "node ../../scripts/cypress.js run --env grepTags=@staging --config specPattern='**/before.cy.ts'" ,
+    "cypress:run:tags:describe": "node ../../scripts/cypress.js run --env grepTags=@smoke --config specPattern='**/describe-tags.cy.ts'" ,
+    "cypress:run:tags:config": "node ../../scripts/cypress.js run --env grepTags=config --config specPattern='**/config-tags.cy.ts'" ,
+    "cypress:run:burn": "node ../../scripts/cypress.js run --env burn=5 --config specPattern='**/burn.cy.ts'" ,
+    "cypress:run:tags:inherit": "node ../../scripts/cypress.js run --env grepTags=@screen-b --config specPattern='**/inherits-tag.cy.ts'" ,
+    "cypress:run:multiple-registrations": "node ../../scripts/cypress.js run --config specPattern='**/multiple-registrations.cy.ts'" ,
+    "cypress:run:tags:nested-describe": "node ../../scripts/cypress.js run --env grepTags='@smoke @integration' --config specPattern='**/nested-describe.cy.ts'" ,
+    "cypress:run:skip": "node ../../scripts/cypress.js run --config specPattern='**/*skip.cy.ts'" ,
+    "cypress:run:this": "node ../../scripts/cypress.js run --env grep='this context' --config specPattern='**/this.cy.ts'" ,
+    "cypress:run:tags": "node ../../scripts/cypress.js run --env grepTags=smoke --config specPattern='**/tags/*.cy.ts'" ,
+    "cypress:run:each": "node ../../scripts/cypress.js run --config specPattern='**/each.cy.ts'" ,
+    "cypress:run:grep": "node ../../scripts/cypress.js run --config specPattern='**/grep-task.cy.ts'" ,
+    "cypress:run:tags:or": "node ../../scripts/cypress.js run --env grepTags='regression high' --config specPattern='**/tags/*.cy.ts'" ,
+    "cypress:run:tags:and": "node ../../scripts/cypress.js run --env grepTags='smoke+high' --config specPattern='**/tags/*.cy.ts'" ,
+    "cypress:run:tags:and:not": "node ../../scripts/cypress.js run --env grepTags='smoke+-high' --config specPattern='**/tags/*.cy.ts'",
+    "cypress:run:or": "node ../../scripts/cypress.js run --env grep='Test 1',grepTags='regression high' --config specPattern='**/tags/*.cy.ts'" ,
+    "cypress:run:and": "node ../../scripts/cypress.js run --env grep='Test 2',grepTags='smoke+high' --config specPattern='**/tags/*.cy.ts'" ,
+    "cypress:run:and:not": "node ../../scripts/cypress.js run --env grep='Test 2',grepTags='smoke+-high' --config specPattern='**/tags/*.cy.ts'",
+    "cypress:run:filter:specs": "node ../../scripts/cypress.js run --env grep='Test 2',grepFilterSpecs=true --config specPattern='**/tags/*.cy.ts'" ,
+    "cypress:run:tags:filter:specs": "node ../../scripts/cypress.js run --env grepTags='regression',grepFilterSpecs=true --config specPattern='**/tags/*.cy.ts'" ,
+    "cypress:run:omit:specs": "node ../../scripts/cypress.js run --env grep='Test 2',grepOmitFiltered=true --config specPattern='**/tags/*.cy.ts'" ,
+    "cypress:run:tags:omit:specs": "node ../../scripts/cypress.js run --env grepTags='regression',grepOmitFiltered=true --config specPattern='**/tags/*.cy.ts'" ,
+    "cypress:run:untagged": "node ../../scripts/cypress.js run --env grepUntagged=true --config specPattern='**/tags/*.cy.ts'" ,
+    "cypress:run:inverted": "node ../../scripts/cypress.js run --env grep='-Test 2' --config specPattern='**/tags/*.cy.ts'" ,
+    "cypress:run:tag:inverted": "node ../../scripts/cypress.js run --env grepTags='-regression' --config specPattern='**/tags/*.cy.ts'" ,
+    "test": "vitest",
+    "test-debug": "vitest --inspect-brk --no-file-parallelism --test-timeout=0",
     "lint": "eslint"
   },
   "dependencies": {
@@ -18,15 +44,25 @@
     "cypress-each": "^1.11.0",
     "cypress-expect": "^2.5.3",
     "eslint": "^9.31.0",
+    "vitest": "^3.2.4",
     "typescript": "~5.4.5"
   },
   "peerDependencies": {
     "cypress": ">=10"
   },
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    },
+    "./plugin": {
+      "types": "./dist/plugin.d.ts",
+      "default": "./dist/plugin.js"
+    }
+  },
   "files": [
-    "src"
+    "dist"
   ],
-  "types": "src/index.d.ts",
   "license": "MIT",
   "repository": {
     "type": "git",
